requirements.txt
python-telegram-bot>=20.0
anthropic
python-dotenv
airtable-python-wrapper 
nest-asyncio